+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Prompt Guard: `classify` and `threshold` parameters.
+
 ## 5.5.0beta2 - 2025-01-30
 
 ### Changed